requirements.txt
speedtest-cli
scapy
dotenv
requests
schedule
pyyaml
psutil